fix: cqrs 엄격한 분리
-- 일반 조회랑 검색 조회 분리
