+++ src/main/java/com/jangburich/application/prepay/service/PrepayQueryService.java
@@ -1,8 +1,5 @@
 package com.jangburich.application.prepay.service;
 
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
 import java.util.List;
 
 import com.jangburich.application.store.resolver.StoreResolver;
@@ -17,7 +14,6 @@ import com.jangburich.global.error.DefaultException;
 import com.jangburich.global.error.DefaultNullPointerException;
 import com.jangburich.global.payload.ErrorCode;
 import com.jangburich.infrastructure.repository.UserRepository;
-import com.jangburich.infrastructure.repository.queryDsl.PointTransactionQueryDslRepository;
 import com.jangburich.presentation.prepay.dto.response.PrepayResponse;
 import com.jangburich.presentation.prepay.dto.response.PrepaymentInfoResponse;
 
@@ -36,7 +32,6 @@ public class PrepayQueryService {
     private final StoreTeamRepository storeTeamRepository;
     private final UserRepository userRepository;
     private final PointTransactionRepository pointTransactionRepository;
-    private final PointTransactionQueryDslRepository pointTransactionQueryDslRepository;
 
     private final StoreResolver storeResolver;
 
@@ -71,25 +66,6 @@ public class PrepayQueryService {
         List<PointTransaction> storePointTransactionList = pointTransactionRepository.findAllByStoreIdOrderByIdDesc(store.getId());
 
         return buildStorePrepayInfoResponse(storePointTransactionList);
-    }
-
-    public List<PrepayResponse.StorePrepayInfo> getSearchStorePrepayInfoByName(String userId, String name) {
-        Store store = storeResolver.getStoreByUserId(userId);
-
-        return pointTransactionQueryDslRepository.queryAllByStoreIdOrderByIdDesc(store.getId(), name);
-    }
-
-    public List<PrepayResponse.StorePrepayInfo> getPrepayInfoFilterByDate(String userId, String startDate, String endDate) {
-        Store store = storeResolver.getStoreByUserId(userId);
-
-        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy.MM.dd");
-        LocalDateTime startDateTime = LocalDate.parse(startDate, formatter).atStartOfDay();
-        LocalDateTime endDateTime = LocalDate.parse(endDate, formatter).atTime(23, 59, 59);
-
-        if (startDateTime.isAfter(endDateTime))
-            throw new IllegalArgumentException("시작날짜는 종료날짜보다 길 수 없습니다.");
-
-        return pointTransactionQueryDslRepository.queryAllByStoreIdAndBetweenStartDateAndEndDateOrderByIdDesc(store.getId(), startDateTime, endDateTime);
     }
 
     private List<PrepayResponse.StorePrepayInfo> buildStorePrepayInfoResponse(List<PointTransaction> storeList) {
